raaaaaaaaaaaaaaaaaaah

+++ code/trame.c
@@ -1,6 +1,6 @@
 #include <trame.h>
 
-void init_trame(trame * tra,mac source,mac dest,int message){
+void init_trame(trame * tra, mac source, mac dest, size_t message){
     tra->dest =dest;
     tra->source = source;
     tra->message =message;
