+++ code/trame.h
@@ -1,11 +1,12 @@
 #pragma once 
 #include <adress.h>
+#include <stddef.h>
 
 typedef struct trame{
     mac source;
     mac dest;
-    int message;
+    size_t message;
 }trame;
 
-void init_trame(trame * tra,mac source,mac dest,int message);
+void init_trame(trame * tra, mac source, mac dest, size_t message);
